+++ .env.docker
@@ -1,4 +1,4 @@
 PORT=3000
 CHATWORK_API_TOKEN=chatwork_api_token_example
-SECRET_TOKEN=github_token_example
+SECRET_TOKEN=iloveyou
 DATABASE_URL=postgres://postgres:password@db:5432/pullnotifier
