+++ .lintstagedrc.json
@@ -1,5 +1,3 @@
 {
-  "*.ts": [
-    "npm run lint"
-  ]
+  "*.ts": ["npm run lint"]
 }
