+++ nodemon.debug.json
@@ -1,14 +1,7 @@
 {
-  "watch": [
-    "src",
-    "static",
-    ".env"
-  ],
+  "watch": ["src", "static", ".env"],
   "verbose": true,
   "ext": "js,ts,json,pr",
-  "ignore": [
-    "src/**/*.spec.ts",
-    "src/**/*.test.ts"
-  ],
+  "ignore": ["src/**/*.spec.ts", "src/**/*.test.ts"],
   "exec": "node --inspect=0.0.0.0:9229 -r ts-node/register -r tsconfig-paths/register src/server.ts"
 }
