+++ src/middlewares/webhook.middleware.ts
@@ -2,16 +2,7 @@ import { RequestHandler } from 'express';
 import { HttpException } from '@exceptions/HttpException';
 import { StatusCodes } from '@utils/status-code';
 import { GITHUB_EVENT, GITHUB_HEADER } from '@/constants/headers';
-import { logger } from '@utils/logger';
-import * as crypto from 'crypto';
 import { MESSAGES } from '@/constants/messages';
-import config from 'config';
-
-const getSignature = buf => {
-  const hmac = crypto.createHmac('sha256', process.env.SECRET_TOKEN);
-  hmac.update(buf, 'utf-8');
-  return 'sha256=' + hmac.digest('hex');
-};
 
 export const githubEventValidation = (req, res, next): RequestHandler => {
   const event = req.header(GITHUB_HEADER.X_GITHUB_EVENT);
@@ -24,6 +15,7 @@ export const githubEventValidation = (req, res, next): RequestHandler => {
     next(new HttpException(StatusCodes.OK, MESSAGES.CALL_GITHUB_WEBHOOK_NO_ACTION));
     return;
   }
+
   next();
 };
 
@@ -36,27 +28,4 @@ export const processGithubData = (req, res, next) => {
   req.body.event = req.header(GITHUB_HEADER.X_GITHUB_EVENT);
 
   next();
-};
-
-export const githubSignatureValidation = (req, res, next) => {
-  try {
-    if (config.get('env') === 'development' || config.get('env') === 'production') {
-      next();
-      return;
-    }
-    const expected = req.header(GITHUB_HEADER.X_HUB_SIGNATURE_256);
-    const payload = req.body.rawBody;
-    const calculated = getSignature(payload);
-    console.log(calculated);
-    logger.info(`${GITHUB_HEADER.X_HUB_SIGNATURE_256}:`, expected, 'Content:', '-' + payload.toString('utf8') + '-');
-    if (expected !== calculated) {
-      next(new HttpException(StatusCodes.UNAUTHORIZED, 'Invalid github webhook signature'));
-    } else {
-      logger.info('Valid signature!');
-      next();
-    }
-  } catch (err) {
-    console.log(err);
-    next(new HttpException(StatusCodes.UNAUTHORIZED, 'Invalid github webhook signature'));
-  }
 };
